feat: Add a enum "FlagTypes"
Originally, parseFlagVal() requires the user to set the type of the
flag by specifying its name as a string. This is error-prone and
inconvenient (one has to memorize the names). This commit
adds a exported variable "FlagTypes", which hides the names
inside it, for example, the user can use `FlagTypes.string`
instead of `"string"`.

The change is backward-compatible.